wireguard: queueing: do not account for pfmemalloc when clearing skb header
commit 04d2ea92a18417619182cbb79063f154892b0150 upstream.

Before 8b7008620b84 ("net: Don't copy pfmemalloc flag in __copy_skb_
header()"), the pfmemalloc flag used to be between headers_start and
headers_end, which is a region we clear when preparing the packet for
encryption/decryption. This is a parameter we certainly want to
preserve, which is why 8b7008620b84 moved it out of there. The code here
was written in a world before 8b7008620b84, though, where we had to
manually account for it. This commit brings things up to speed.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: Ieafbf9fa8c569aa728725e8211cde732ef59f27e
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: Ieafbf9fa8c569aa728725e8211cde732ef59f27e
(cherry picked from commit ae5ae94a2eb93d61ae7787fb1c68193034aa5a37)
Signed-off-by: TogoFire <togofire@mailfence.com>